Extract reporting batch contract facade
